+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- add limit for staking
+- add requirement for more than one signature
+
+### Changed
+- merge `StakingBank` and `ValidatorRegistry`
 
 ## [2.0.7] - 2021-07-13
 ### Fixed